Fixed GC bug, stamina bug, and more balancing

+++ java/lorien/legacies/legacies/implementations/AccelixLegacy.java
@@ -26,7 +26,7 @@ public class AccelixLegacy extends Legacy
 	}
 
 	@Override
-	public int getStaminaPerTick()
+	public float getStaminaPerTick()
 	{
 		return toggled ? 10 : 0;
 	}
