+++ java/lorien/legacies/legacies/implementations/NovisLegacy.java
@@ -45,7 +45,7 @@ public class NovisLegacy extends Legacy {
 	}
 	
 	@Override
-	public int getStaminaPerTick()
+	public float getStaminaPerTick()
 	{
 		return toggled ? 10 : 0;
 	}
