+++ java/lorien/legacies/legacies/implementations/NoxenLegacy.java
@@ -49,7 +49,7 @@ public class NoxenLegacy extends Legacy {
 	}
 	
 	@Override
-	public int getStaminaPerTick()
+	public float getStaminaPerTick()
 	{
 		return 0;
 	}
