+++ java/lorien/legacies/legacies/implementations/Telekinesis.java
@@ -322,7 +322,7 @@ public class Telekinesis extends Legacy
 	}
 
 	@Override
-	public int getStaminaPerTick()
+	public float getStaminaPerTick()
 	{
 		return 0;
 	}
